Add support for SCHEMAS env var
This is a bit of a hack to address an issue with local relative file references.

e.g.

* Place
* Person references PostalAddress object in Place
* Organization references Person

In this case, Person will instatiate correctly, but Organization fails when it tries to load Person as the path to the schema file for the Place object is wrong (the assumed base path is incorrect for nested dependancies).

I am inenlagently, but simply, solving it by using a SCHEMAS environment variable, whcih is used as the root path for all reference to local schema files.