resolve-external: keep track of original paths from references
Right now the library is “loosing” the original value
information from the `$ref` entry in case of we are resolving relative
file paths.

It can be a useful information to keep the original path taken from
the JSON Schema document while crawling the references tree. Thus this
PR is trying to keep a `Set` containing all original paths when
resolving external $ref pointers.