Fix tests when run in paths with spaces
Previously, on Windows and Linux (comment here was incorrect) these
tests failed if the path contained spaces, because the err.source
property used %20 instead of a literal space. This happens because
internal path representations are used, instead of the actual path.

This change ensures that the path representations are at least decoded
correctly before any errors are thrown, ensuring that usable paths are
returned. It is possible to go further and use url.toFileSystemPath to
fully transform the URL-formatted path into a standard platform-specific
filesystem path, but this could be a breaking change for some use cases,
because it would change from unix to Windows paths on Windows, for
example.