chore: Improve DOM-less type exports
Usage of HeadersInit requires either a node-fetch dependency or to
include DOM in tsconfig libs for consumer projects. It is not exported
as a global type in @types/node v18 or v20. RequestInit is:

https://github.com/DefinitelyTyped/DefinitelyTyped/blob/aaf94a0a6431f52c46dab26182f1156d79720d29/types/node/v18/globals.d.ts#L341

Define headers in interface HTTPResolverOptions<> via
RequestInit['headers'] which will make it possible to depend on this
project without node-fetch.